+++ tests/test_photo.py
@@ -1,0 +1,54 @@
+"""
+Unit tests for the Photo class
+
+"""
+
+from __future__ import print_function, division
+
+import os
+import unittest
+
+from hugophotoswipe.photo import Photo
+from hugophotoswipe.conf import settings
+
+class PhotoTestCase(unittest.TestCase):
+
+    def setUp(self):
+        pth = os.path.realpath(__file__)
+        dr = os.path.dirname(pth)
+        tst = os.path.join(dr, 'test.jpg')
+        self.photo = Photo(album_name='test_album', original_path=tst, 
+                name='test_image', alt='Alt text', caption='caption text', 
+                copyright=None)
+
+    def test_resize_dims(self):
+        """ [Photo]: Test resize dimensions """
+
+        # testing all possible modes
+        pairs = [
+                ('dim_max_large', 'large'),
+                ('dim_max_small', 'small'),
+                ('dim_max_thumb', 'thumb'),
+                ('dim_max_cover', 'cover')
+                ]
+        for setting_name, mode in pairs:
+            setattr(settings, setting_name, '1600')
+            #settings.dim_max_large = '1600'
+            dims = self.photo.resize_dims(mode)
+            # 1040 = 1600 / 1800 * 1170
+            self.assertEqual(dims, (1600, 1040))
+
+            setattr(settings, setting_name, '1500x')
+            dims = self.photo.resize_dims(mode)
+            # 1500 / 1800 * 1170 = 975
+            self.assertEqual(dims, (1500, 975))
+
+            setattr(settings, setting_name, 'x1000')
+            dims = self.photo.resize_dims(mode)
+            # 1000 / 1170 * 1800 = 1538
+            self.assertEqual(dims, (1538, 1000))
+
+            setattr(settings, setting_name, '800x600')
+            dims = self.photo.resize_dims(mode)
+            self.assertEqual(dims, (800, 600))
+